fixed atlas example to use 0.24 of image

+++ examples/atlas/main.rs
@@ -72,8 +72,8 @@ fn main() {
         image::imageops::replace(
             &mut output,
             &img.image,
-            img.location.x as u32,
-            img.location.y as u32,
+            img.location.x as i64,
+            img.location.y as i64,
         );
     }
     let output_file = basedir.join("atlas.png");
